% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_Vertices.R
\name{assess.Vertices}
\alias{assess.Vertices}
\title{Set up the data for the NM algorithm}
\usage{
assess.Vertices(vertex.list, in.data)
}
\description{
Takes the input values, calcs StdDevs and means, then
}
