% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_Evolution.R
\name{differential.Evolution}
\alias{differential.Evolution}
\title{Differential evolution for discrete data with finite support.}
\usage{
differential.Evolution(in_counts, contenter_chng_pts, n_gen = 1e+06)
}
\description{
FIXME
}
