% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_LR.R
\name{evaluate_LR}
\alias{evaluate_LR}
\title{evaluate the log of the skellam likelihood ratio}
\usage{
evaluate_LR(clan_avg, right_avg, left_avg, clan_diff)
}
\description{
A function to directly calcualte the likelihood ratio for the skellam distribution
}
