% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skellam_LR_bound.R
\name{skellam_LR_bound}
\alias{skellam_LR_bound}
\title{Bound on the skellam likelihood ratio.}
\usage{
skellam_LR_bound(w = NULL, left_avg = NULL, right_avg = NULL,
  null_avg = NULL, lower_bound = TRUE)
}
\arguments{
\item{w}{A numeric of length 1.  The osverved value of the Skellam random variable.}

\item{left_avg}{The estimate of the intensity parameter for the observations within
the left group.}

\item{right_avg}{The estimate of the intensity parameter for the observations within
the right group.}

\item{null_avg}{The estimate of the intensity parameter for the observations within
the combined left and right groups.}

\item{lower_bound}{Logical.  Determines whether to calculate the lower of the upper bounds of 
the skellam likelihood ratio.}
}
\description{
Because the modiffied bessel function fails for large indices, this
    function applied the findings of
}
\references{
\insertRef{balachandran2013exponential}{changePOIssoNT}
}
