% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skellamLR.R
\name{skellamLR}
\alias{skellamLR}
\title{Skellam Likelihood ratio for varying group sizes}
\usage{
skellamLR(contender_chng_pt, left_and_right_reaches, in_counts)
}
\arguments{
\item{left_and_right_reaches}{A dataframe containing the candidate 
changepoints along with the maximum group sizes separated by their 
respective changepoint.}

\item{in_counts}{A dataframe containing the observed counts}
}
\value{
numeric
}
\description{
Takes in a data frame of candidate changepoints and clique 
    capacities, then calculates the likelihood ratio for the difference
    of each clique.
}
