% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skellamLR.R
\name{skellamLR}
\alias{skellamLR}
\title{Skellam Likelihood ratio for varying window sizes}
\usage{
skellamLR(in.data, windows)
}
\arguments{
\item{in.data}{a dataframe containing the observed counts}

\item{windows}{A dataframe containing the candidate changepoints along with the maximum possible values for each window}
}
\description{
Takes in a data frame of changepoints and window sizes, then calculates the LR for each
}
